Add API State to Country Validation
Spent around an hour this morning before work to pick an item off the
todo list. The only reason it took so long was I misread the
instructions and had the host name swap backwards and was convinced
services was still in there shame. There is one thing of note in this
code and I honestly don't have an answer for it. If you check out
matchStateToCountry.js line 32 you see I throw a custom error. The API
did not error I just did not like the result. It could easily be argued
that that should be a resolve and the logic handled else where.

- Add API state to country validation to isValid function