Add Controllers
I meant to start working but for whatever reason could not close out and
wrote a few more lines of code. I have not run this because I closed
everything accept the editor but I bet it it will work just fine. This
is sort of a half commit then I suppose.

- add edit controller
- add delete controller
- move isValid to pre save hook

Note: the pre save is valid hook should be updated to only call the API
if there is an update to the state or country. I probably wont get
  around to doing this but that was the justification for the move. Pre
  hooks in mongoose are great for just this sort of comparison.